[it] Split ITs for better isolation and scalability
