+++ xmvn-it/src/test/java/org/fedoraproject/xmvn/it/maven/mojo/builddep/ProfilesBuilddepIntegrationTest.java
@@ -1,0 +1,36 @@
+/*-
+ * Copyright (c) 2015-2021 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.fedoraproject.xmvn.it.maven.mojo.builddep;
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * Integration tests for builddep MOJO.
+ * 
+ * @author Mikolaj Izdebski
+ */
+public class ProfilesBuilddepIntegrationTest
+    extends AbstractBuilddepIntegrationTest
+{
+    @Test
+    public void testBuilddepProfiles()
+        throws Exception
+    {
+        expectBuildDependency( "junit", "junit" );
+        expectBuildDependency( "org.codehaus.plexus", "plexus-component-metadata" );
+        performBuilddepTest();
+    }
+}
