+++ xmvn-it/src/test/java/org/fedoraproject/xmvn/it/maven/mojo/javadoc/JavadocJPMSIntegrationTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fedoraproject.xmvn.it.maven.mojo;
+package org.fedoraproject.xmvn.it.maven.mojo.javadoc;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -22,26 +22,16 @@ import java.nio.file.Paths;
 
 import org.junit.jupiter.api.Test;
 
+import org.fedoraproject.xmvn.it.maven.mojo.AbstractMojoIntegrationTest;
+
 /**
  * Integration tests for javadoc MOJO.
  * 
  * @author Mikolaj Izdebski
  */
-public class JavadocIntegrationTest
+public class JavadocJPMSIntegrationTest
     extends AbstractMojoIntegrationTest
 {
-    @Test
-    public void testJavadoc()
-        throws Exception
-    {
-        performTest( "org.fedoraproject.xmvn:xmvn-mojo:javadoc" );
-
-        assertTrue( Files.isDirectory( Paths.get( "target/xmvn-apidocs" ) ) );
-        assertTrue( Files.isRegularFile( Paths.get( "target/xmvn-apidocs/foo/Bar.html" ) ) );
-        assertTrue( Files.isSymbolicLink( Paths.get( ".xmvn/apidocs" ) ) );
-        assertTrue( Files.isSameFile( Paths.get( ".xmvn/apidocs" ), Paths.get( "target/xmvn-apidocs" ) ) );
-    }
-
     @Test
     public void testJavadocJPMS()
         throws Exception
