+++ xmvn-it/src/test/resources/testBuilddepTychoPomless/.mvn/extensions.xml
@@ -3,6 +3,6 @@
   <extension>
     <groupId>org.eclipse.tycho.extras</groupId>
     <artifactId>tycho-pomless</artifactId>
-    <version>0.25.0</version>
+    <version>2.3.0</version>
   </extension>
 </extensions>
