+++ xmvn-it/src/test/resources/testBuilddepTychoPomless/pom.xml
@@ -13,7 +13,7 @@
       <plugin>
 	<groupId>org.eclipse.tycho</groupId>
 	<artifactId>tycho-maven-plugin</artifactId>
-	<version>1.5.1</version>
+	<version>2.3.0</version>
 	<extensions>true</extensions>
       </plugin>
     </plugins>
